+++ bench/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "rust-sitter-example"
+name = "rust-sitter-benchmarking"
 version = "0.4.4"
 authors = ["Shadaj Laddad <shadaj@users.noreply.github.com>"]
 edition = "2021"
@@ -12,12 +12,12 @@ tree-sitter-standard = ["rust-sitter/tree-sitter-standard"]
 
 [dependencies]
 rust-sitter = { path = "../runtime", default-features = false }
-codemap = "0.1.3"
-codemap-diagnostic = "0.1.1"
+rand = "0.9"
+criterion = { version = "0.5", features = ["html_reports"] }
 
 [build-dependencies]
 rust-sitter-tool = { path = "../tool" }
 
-[dev-dependencies]
-insta = "1.39"
-wasm-bindgen-test = "0.3.0"
+[[bench]]
+name = "arena_benchmark"
+harness = false
