--- example/src/lib.rs
@@ -1,5 +1,0 @@
-pub mod arithmetic; // Used for benchmark
-mod handles;
-mod optionals;
-mod repetitions;
-mod words;
